requirements.txt
# Core Flask & Web
Flask==3.1.2
python-dotenv==1.1.1

# Vector Store
faiss-cpu==1.12.0

# PDF Processing
pypdf==6.1.1

# LangChain Core Packages
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-text-splitters==0.3.11

# AI Integrations
langchain-google-genai==2.1.12
langchain-groq==0.3.8

# Google AI
google-generativeai==0.3.2
google-api-core==2.11.0
google-auth==2.29.0

# Utilities - Use versions with pre-compiled wheels
numpy==1.26.4; python_version<"3.13"
numpy==2.0.0; python_version>="3.13"
requests==2.32.5
tqdm==4.67.1
protobuf==4.25.3

# Development
setuptools==69.0.3
wheel==0.45.1